Cleanup some of the dead code annotations. (#379)
