+++ src/test_utils.rs
@@ -1,13 +1,10 @@
 use crate::config::Clock;
 use crate::iter::KeyValueIterator;
 use crate::types::{KeyValue, RowAttributes, ValueDeletable};
-use bytes::{BufMut, Bytes, BytesMut};
+use bytes::Bytes;
 use rand::Rng;
 use std::sync::atomic::{AtomicI64, Ordering};
 
-// this complains because we include these in the bencher feature but they are only
-// used for cfg(test)
-#[allow(dead_code)]
 pub(crate) async fn assert_iterator<T: KeyValueIterator>(
     iterator: &mut T,
     entries: &[(Vec<u8>, ValueDeletable, RowAttributes)],
@@ -49,15 +46,11 @@ pub(crate) async fn assert_iterator<T: KeyValueIterator>(
         .is_none());
 }
 
-// this complains because we include these in the bencher feature but they are only
-// used for cfg(test)
-#[allow(dead_code)]
 pub fn assert_kv(kv: &KeyValue, key: &[u8], val: &[u8]) {
     assert_eq!(kv.key, key);
     assert_eq!(kv.value, val);
 }
 
-#[allow(dead_code)]
 pub(crate) fn gen_attrs(ts: i64) -> RowAttributes {
     RowAttributes {
         ts: Some(ts),
@@ -65,7 +58,6 @@ pub(crate) fn gen_attrs(ts: i64) -> RowAttributes {
     }
 }
 
-#[allow(dead_code)]
 pub(crate) fn gen_empty_attrs() -> RowAttributes {
     RowAttributes {
         ts: None,
@@ -77,7 +69,6 @@ pub(crate) struct TestClock {
     pub(crate) ticker: AtomicI64,
 }
 
-#[allow(dead_code)]
 impl TestClock {
     pub(crate) fn new() -> TestClock {
         TestClock {
@@ -92,56 +83,6 @@ impl Clock for TestClock {
     }
 }
 
-#[derive(Clone)]
-pub(crate) struct OrderedBytesGenerator {
-    suffix: Bytes,
-    bytes: Vec<u8>,
-    min: u8,
-    max: u8,
-}
-
-impl OrderedBytesGenerator {
-    #[allow(dead_code)]
-    pub(crate) fn new_with_suffix(suffix: &[u8], bytes: &[u8]) -> Self {
-        Self::new(suffix, bytes, u8::MIN, u8::MAX)
-    }
-
-    // this complains because we include these in the bencher feature but they are only
-    // used for cfg(test)
-    #[allow(dead_code)]
-    pub(crate) fn new_with_byte_range(bytes: &[u8], min: u8, max: u8) -> Self {
-        Self::new(&[], bytes, min, max)
-    }
-
-    pub(crate) fn new(suffix: &[u8], bytes: &[u8], min: u8, max: u8) -> Self {
-        let bytes = Vec::from(bytes);
-        Self {
-            suffix: Bytes::copy_from_slice(suffix),
-            bytes,
-            min,
-            max,
-        }
-    }
-
-    pub(crate) fn next(&mut self) -> Bytes {
-        let mut result = BytesMut::with_capacity(self.bytes.len() + std::mem::size_of::<u32>());
-        result.put_slice(self.bytes.as_slice());
-        result.put(self.suffix.as_ref());
-        self.increment();
-        result.freeze()
-    }
-
-    fn increment(&mut self) {
-        let mut pos = self.bytes.len() - 1;
-        while self.bytes[pos] == self.max {
-            self.bytes[pos] = self.min;
-            pos -= 1;
-        }
-        self.bytes[pos] += 1;
-    }
-}
-
-#[allow(dead_code)]
 pub(crate) fn gen_rand_bytes(n: usize) -> Bytes {
     let mut rng = rand::thread_rng();
     let random_bytes: Vec<u8> = (0..n).map(|_| rng.gen()).collect();
@@ -150,7 +91,6 @@ pub(crate) fn gen_rand_bytes(n: usize) -> Bytes {
 
 // it seems that insta still does not allow to customize the snapshot path in insta.yaml,
 // we can remove this macro once insta supports it.
-#[cfg(test)]
 macro_rules! assert_debug_snapshot {
     ($name:expr, $output:expr) => {
         let mut settings = insta::Settings::clone_current();
@@ -160,52 +100,4 @@ macro_rules! assert_debug_snapshot {
     };
 }
 
-#[cfg(test)]
 pub(crate) use assert_debug_snapshot;
-
-#[cfg(test)]
-mod tests {
-    use bytes::{BufMut, Bytes};
-
-    use crate::test_utils::OrderedBytesGenerator;
-
-    #[test]
-    fn test_should_generate_ordered_bytes() {
-        let mut suffix = Vec::<u8>::new();
-        suffix.put_u32(3735928559);
-        let start = [0u8, 0u8, 0u8];
-        let mut gen = OrderedBytesGenerator::new(suffix.as_ref(), &start, 0, 2);
-
-        let expected = [
-            [0u8, 0u8, 0u8, 0xde, 0xad, 0xbe, 0xef],
-            [0u8, 0u8, 1u8, 0xde, 0xad, 0xbe, 0xef],
-            [0u8, 0u8, 2u8, 0xde, 0xad, 0xbe, 0xef],
-            [0u8, 1u8, 0u8, 0xde, 0xad, 0xbe, 0xef],
-            [0u8, 1u8, 1u8, 0xde, 0xad, 0xbe, 0xef],
-            [0u8, 1u8, 2u8, 0xde, 0xad, 0xbe, 0xef],
-            [0u8, 2u8, 0u8, 0xde, 0xad, 0xbe, 0xef],
-            [0u8, 2u8, 1u8, 0xde, 0xad, 0xbe, 0xef],
-            [0u8, 2u8, 2u8, 0xde, 0xad, 0xbe, 0xef],
-            [1u8, 0u8, 0u8, 0xde, 0xad, 0xbe, 0xef],
-            [1u8, 0u8, 1u8, 0xde, 0xad, 0xbe, 0xef],
-            [1u8, 0u8, 2u8, 0xde, 0xad, 0xbe, 0xef],
-            [1u8, 1u8, 0u8, 0xde, 0xad, 0xbe, 0xef],
-            [1u8, 1u8, 1u8, 0xde, 0xad, 0xbe, 0xef],
-            [1u8, 1u8, 2u8, 0xde, 0xad, 0xbe, 0xef],
-            [1u8, 2u8, 0u8, 0xde, 0xad, 0xbe, 0xef],
-            [1u8, 2u8, 1u8, 0xde, 0xad, 0xbe, 0xef],
-            [1u8, 2u8, 2u8, 0xde, 0xad, 0xbe, 0xef],
-            [2u8, 0u8, 0u8, 0xde, 0xad, 0xbe, 0xef],
-            [2u8, 0u8, 1u8, 0xde, 0xad, 0xbe, 0xef],
-            [2u8, 0u8, 2u8, 0xde, 0xad, 0xbe, 0xef],
-            [2u8, 1u8, 0u8, 0xde, 0xad, 0xbe, 0xef],
-            [2u8, 1u8, 1u8, 0xde, 0xad, 0xbe, 0xef],
-            [2u8, 1u8, 2u8, 0xde, 0xad, 0xbe, 0xef],
-            [2u8, 2u8, 0u8, 0xde, 0xad, 0xbe, 0xef],
-            [2u8, 2u8, 1u8, 0xde, 0xad, 0xbe, 0xef],
-        ];
-        for e in expected.iter() {
-            assert_eq!(gen.next(), Bytes::copy_from_slice(e))
-        }
-    }
-}
